zeroguard_idp/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
flask-limiter==3.7.0
python-dotenv==1.0.1
bcrypt==4.1.3
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==42.0.8
PyJWT==2.8.0
Pillow==10.3.0
requests==2.31.0
